requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.0
